Move health check tests into its own file
